WIP cpuinfo.max_freq is set to the highest frequency in cpufreq_frequency_table (defined in tegra20-cpufreq.c)
cpuinfo.max_freq value can be used by cpufreq scaling governors
(e.g. sched and schedutil).

The tegra20 cpufreq driver includes the overclocked speeds in the
cpufreq frequency table. A result is that the governor will assume
the stock frequency of CPU is the max overclock frequency.
The scaling algorithms will be based on the max overclock frequency.