DT: adjust jack zone detect delay
The downstream stmpe811 adc driver had a large delay for reading adc
register. The mainline stmpe811 adc driver uses a microsecond delay which
is closer to the data sheet.

Compensate for the difference by increasing the delay in the
jack driver. This is required to get proper sample values
for headphone jack detection.