clk: tegra: Convert CCLKG mux to mux + clock divider on Tegra30
We want to use the common CPUFreq driver for both Tegra20 and Tegra30.
In order to make a common driver without much churning, let's restructure
the CCLKG clock definition such that it will have a PLLP parent like
Tegra20. That means that we want to convert CCLKG from a MUX to a proper
MUX + divider, removing the custom pll_p_cclkg (and alike) virtual clocks.

Signed-off-by: Dmitry Osipenko <digetx@gmail.com>